+++ CHANGELOG.md
@@ -1,14 +1,31 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2018-12-20
+## [0.0.2] - 2018-12-27
+
+### Added
+
+-   Button tab on project grid
+-   Active/non-active button status on project grid identicating the status
+
 ### Changed
-- Header title to "open-sans"
+
+-   logo font
+-   Switched "Project" and "creator" at project grid
+-   sizing of the logo and font
+
+## [0.0.1] - 2018-12-20
+
+### Changed
+
+-   Header title to "open-sans"
 
 ## Added
-- Changelog
 
-[0.0.1]: 
+-   Changelog
+
+[0.0.1]: